Rst export fixes (#996)
* Adding WeasyPrint last Python2 version

If you build projecta from scratch, pip will fail because latest
WeasyPrint is Python3 only. So adding the latest Python2 version
number in the REQUIREMENTS.txt fixes this.

* Removing Sponsor table from RST export

The (by Pandoc) generated table rst has some many troubles compiling
in sphinx that it is better to not add it here.
In newer version of qgis.org changelogs the sponsors will be inserted
via runtime javascript injections

* Use real images instead of thumbnails for rst export

Though my preference would be to have clickable thumbnails in the
rst/html current setup was not working.
So back to full images for now.